requirements.txt
numpy
librosa
sounddevice
scipy
scikit-learn
speechbrain
torch
torchaudio
noisereduce
pyyaml